+++ ts/features/wallet/component/PaymentMethodFeatures.tsx
@@ -1,23 +1,11 @@
-import { View } from "native-base";
 import * as React from "react";
-import { StyleSheet } from "react-native";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
-import { BpdConfig } from "../../../../definitions/content/BpdConfig";
-import { H3 } from "../../../components/core/typography/H3";
-import { IOColors } from "../../../components/core/variables/IOColors";
-import ItemSeparatorComponent from "../../../components/ItemSeparatorComponent";
-import IconFont from "../../../components/ui/IconFont";
-import { bpdEnabled } from "../../../config";
-import I18n from "../../../i18n";
 import { bpdRemoteConfigSelector } from "../../../store/reducers/backendStatus";
 import { GlobalState } from "../../../store/reducers/types";
-import {
-  EnableableFunctionsTypeEnum,
-  PaymentMethod
-} from "../../../types/pagopa";
-import BpdPaymentMethodCapability from "../../bonus/bpd/components/BpdPaymentMethodCapability";
-import PagoPaPaymentCapability from "./PagoPaPaymentCapability";
+import { PaymentMethod } from "../../../types/pagopa";
+import PaymentMethodInitiatives from "./PaymentMethodInitiatives";
+import PaymentMethodSettings from "./PaymentMethodSettings";
 
 type OwnProps = { paymentMethod: PaymentMethod };
 
@@ -25,83 +13,19 @@ type Props = ReturnType<typeof mapDispatchToProps> &
   ReturnType<typeof mapStateToProps> &
   OwnProps;
 
-const styles = StyleSheet.create({
-  icon: { alignSelf: "center" },
-  row: { flex: 1, flexDirection: "row" }
-});
-
-const capabilityFactory = (
-  paymentMethod: PaymentMethod,
-  capability: EnableableFunctionsTypeEnum,
-  index: number,
-  bpdRemoteConfig?: BpdConfig
-) => {
-  switch (capability) {
-    case EnableableFunctionsTypeEnum.FA:
-    case EnableableFunctionsTypeEnum.pagoPA:
-      return null;
-    case EnableableFunctionsTypeEnum.BPD:
-      return bpdEnabled && bpdRemoteConfig?.program_active ? (
-        <>
-          <BpdPaymentMethodCapability
-            paymentMethod={paymentMethod}
-            key={`capability_item_${index}`}
-          />
-          <ItemSeparatorComponent noPadded={true} />
-        </>
-      ) : null;
-  }
-};
-
-/** *
- * Extracts the capabilities from a {@link PaymentMethod}, based on the enableableFunctions
- * @param paymentMethod
- * @param bpdRemoteConfig
- */
-const generateCapabilityItems = (
-  paymentMethod: PaymentMethod,
-  bpdRemoteConfig?: BpdConfig
-): ReadonlyArray<React.ReactNode> =>
-  paymentMethod.enableableFunctions.reduce((acc, val, i): ReadonlyArray<
-    React.ReactNode
-  > => {
-    const handlerForCapability = capabilityFactory(
-      paymentMethod,
-      val,
-      i,
-      bpdRemoteConfig
-    );
-    return handlerForCapability === null ? acc : [...acc, handlerForCapability];
-  }, [] as ReadonlyArray<React.ReactNode>);
-
 /**
- * Display the capabilities available for a payment method
+ * Display the features available for a payment method:
+ * - vertical initiatives (eg: cashback, fa)
+ * - global settings (payment capability, favourite, etc.)
  * @param props
  * @constructor
  */
-const PaymentMethodFeatures: React.FunctionComponent<Props> = props => {
-  const capabilityItems = generateCapabilityItems(
-    props.paymentMethod,
-    props.bpdRemoteConfig
-  );
-
-  return (
-    <>
-      <View style={styles.row}>
-        <IconFont
-          name={"io-preferenze"}
-          size={20}
-          color={IOColors.bluegreyDark}
-          style={styles.icon}
-        />
-        <View hspacer={true} />
-        <H3 color={"bluegrey"}>{I18n.t("wallet.capability.title")}</H3>
-      </View>
-      {capabilityItems}
-      <PagoPaPaymentCapability paymentMethod={props.paymentMethod} />
-    </>
-  );
-};
+const PaymentMethodFeatures: React.FunctionComponent<Props> = props => (
+  <>
+    <PaymentMethodInitiatives paymentMethod={props.paymentMethod} />
+    <PaymentMethodSettings paymentMethod={props.paymentMethod} />
+  </>
+);
 
 const mapDispatchToProps = (_: Dispatch) => ({});
