+++ ts/features/wallet/component/PaymentMethodSettings.tsx
@@ -1,0 +1,26 @@
+import * as React from "react";
+import { connect } from "react-redux";
+import { Dispatch } from "redux";
+import { GlobalState } from "../../../store/reducers/types";
+import { PaymentMethod } from "../../../types/pagopa";
+import PagoPaPaymentCapability from "./PagoPaPaymentCapability";
+
+type OwnProps = { paymentMethod: PaymentMethod };
+
+type Props = ReturnType<typeof mapDispatchToProps> &
+  ReturnType<typeof mapStateToProps> &
+  OwnProps;
+
+const PaymentMethodSettings = (props: Props): React.ReactElement => (
+  <>
+    <PagoPaPaymentCapability paymentMethod={props.paymentMethod} />
+  </>
+);
+
+const mapDispatchToProps = (_: Dispatch) => ({});
+const mapStateToProps = (_: GlobalState) => ({});
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(PaymentMethodSettings);
